requirements.txt
Django>=6.0.3
djangorestframework>=3.14.0
drf-yasg>=1.21.7
PyJWT>=2.8.0
